Set `IDecimal` attributes `min`, `max` and `default` as `Decimal` type instead of `Number`.
